fix sth, need to do: readfile

+++ src/svm_struct_api_types.h
@@ -57,7 +57,7 @@ typedef struct pattern {
   /* this defines the x-part of a training example, e.g. the structure
      for storing a natural language sentence in NLP parsing */
   int sentenceLen;
-  int **frames;
+  double **frames;
 } PATTERN;
 
 typedef struct label {
